+++ src/chinese/segmentor/agendaplus.cpp
@@ -236,19 +236,19 @@ SCORE_TYPE getOrUpdateAppendScore(CSegmentor *segmentor, const CStringVector* se
    // about the chars
    static int char_info; // start, middle, end or standalone
    char_info = encodeCharSegmentation(bWordStart, false);
-   const unsigned int end = item->getWordEnd(index);
+   const unsigned long int end = item->getWordEnd(index);
    assert( end<sentence->size()-1 );
 
    retval = 0;
    // ===================================================================================
    // character scores 
-   for (tmp_i = max(0, static_cast<int>(end)-0); tmp_i < min(static_cast<unsigned>(sentence->size()), end+1); ++tmp_i) 
+   for (tmp_i = max(0, static_cast<int>(end)-0); tmp_i < min(static_cast<unsigned long>(sentence->size()), end+1); ++tmp_i) 
       retval += weight.m_mapCharUnigram.getOrUpdateScore( make_pair( ( amount==0 ? segmentor->findWordFromCache(tmp_i,1,sentence)
                                                                                  : segmentor->replaceWordToCache(tmp_i,1,sentence) ),
                                                           encodeCharInfoAndPosition(char_info, tmp_i-end) ), which_score, 
                                                           amount, round);
 
-   for (tmp_i = max(0, static_cast<int>(end)-1); tmp_i < min(static_cast<unsigned>(sentence->size())-1, end+1); ++tmp_i) 
+   for (tmp_i = max(0, static_cast<int>(end)-1); tmp_i < min(static_cast<unsigned long>(sentence->size())-1, end+1); ++tmp_i) 
       retval += weight.m_mapCharBigram.getOrUpdateScore( make_pair( ( amount==0 ? segmentor->findWordFromCache(tmp_i,2,sentence)
                                                                                 : segmentor->replaceWordToCache(tmp_i,2,sentence) ),
                                                          encodeCharInfoAndPosition(char_info, tmp_i-end) ), which_score, 
@@ -380,7 +380,7 @@ void CFeatureHandle::updateScoreVector(const CStringVector* output, const CStrin
  *
  *--------------------------------------------------------------*/
 
-void updateScoreVector(CSegmentor *segmentor, const CStringVector* sentence, unsigned int index, CStateItem* output, CStateItem* correct, int round) {
+void updateScoreVector(CSegmentor *segmentor, const CStringVector* sentence, unsigned long int index, CStateItem* output, CStateItem* correct, int round) {
    static CStateItem temp;
    static int char_index;
    static int word_index;
@@ -462,12 +462,12 @@ bool work(CSegmentor *segmentor, const CStringVector &sentence, CRule &rules, CS
    int m, n;                                    // temporary nums
    int index, temp_index;                       // the index of the current char
    int add_score, subtract_score;               // the score to be subtracted (previous item)
-   static unsigned int doneWordRnd[MAX_SENTENCE_SIZE];  // mask whether candidate with the last word has been cached
-   static unsigned int doneWordLink[MAX_SENTENCE_SIZE]; // link to the corresponding cache state item from word_length + 1
+   static unsigned long int doneWordRnd[MAX_SENTENCE_SIZE];  // mask whether candidate with the last word has been cached
+   static unsigned long int doneWordLink[MAX_SENTENCE_SIZE]; // link to the corresponding cache state item from word_length + 1
    static CStateItem doneWordItems[BEAM_SIZE];          // the allocated cache state items stores the candidates
    static int doneItemPointer; 
    static CStateItem temp_it, candidate;
-   unsigned int word_length;
+   unsigned long int word_length;
    bool bCompatible; 
    const int length = sentence.size();
 
@@ -613,7 +613,7 @@ void CSegmentor::train(const CStringVector* sentence_input, const CStringVector*
    static CStringVector sentence;
    static CRule rules(m_Feature->m_bRule);
    rules.segment(sentence_input, &sentence);
-   const unsigned int length = sentence.size();
+   const unsigned long int length = sentence.size();
 
    assert(length<MAX_SENTENCE_SIZE);
 
@@ -660,7 +660,7 @@ void CSegmentor::segment(const CStringVector* sentence_input, CStringVector *vRe
    static CStringVector sentence;
    static CRule rules(m_Feature->m_bRule); 
    rules.segment(sentence_input, &sentence); 
-   const unsigned int length = sentence.size();
+   const unsigned long int length = sentence.size();
 
    assert(length<MAX_SENTENCE_SIZE);
    assert(vReturn!=NULL);
