+++ src/include/bitarray.h
@@ -22,10 +22,10 @@
 class CBitArray {
    protected:
       char *m_array;
-      unsigned int m_size;
+      unsigned long int m_size;
       int m_slots; // simply for convenience
    public:
-      CBitArray(unsigned int size) : m_size(size) {
+      CBitArray(unsigned long int size) : m_size(size) {
          m_slots = size / 8;
          m_slots += size%8==0 ? 0 : 1;
          m_array = new char[m_slots];
@@ -38,25 +38,25 @@ class CBitArray {
          for (int i=0; i<m_slots; ++i)
             m_array[i] = 0;
       }
-      void set(unsigned int index) {
+      void set(unsigned long int index) {
          assert(index<m_size);
          int slot = index / 8;
          int bit_in_slot = index % 8;
          m_array[slot] |= (1 << bit_in_slot);
       }
-      void unset(unsigned int index) {
+      void unset(unsigned long int index) {
          assert(index<m_size);
          int slot = index / 8;
          int bit_in_slot = index % 8;
          m_array[slot] &= ~(static_cast<char>(1 << bit_in_slot));
       }
-      void flip(unsigned int index) {
+      void flip(unsigned long int index) {
          assert(index<m_size);
          int slot = index / 8;
          int bit_in_slot = index % 8;
          m_array[slot] ^= (1 << bit_in_slot);
       }
-      bool isset(unsigned int index) const {
+      bool isset(unsigned long int index) const {
          assert(index<m_size);
          int slot = index / 8;
          int bit_in_slot = index % 8;
